apps/orchestrator/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
httpx>=0.27.0
openai>=1.30.0
deepgram-sdk>=3.2.0,<4.0.0
twilio>=9.0.0
supabase>=2.4.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
pytz>=2024.1
python-multipart>=0.0.9
